@@ -50,8 +50,8 @@ description = """PCGen"""
 
 defaultTasks 'build'
 
-sourceCompatibility = "1.11"
-targetCompatibility = "1.11"
+sourceCompatibility = "1.16"
+targetCompatibility = "1.16"
 
 ext {
     binDir = "code/bin"
@@ -382,7 +382,7 @@ clean {
 build.mustRunAfter clean
 
 task downloadJRE {
-    def major = 14
+    def major = 16
     def archs = ['x64','x32']
     def osList = ['windows', 'mac','linux']
     def extension = 'zip'
@@ -463,7 +463,7 @@ task downloadJRE {
 
 
 task downloadJavaFXModules(dependsOn: downloadJRE) {
-    def major = "14-0-1"
+    def major = "16-0-0"
     def archs = ['x64', 'x86']
     def osList = ["windows", "mac", "linux"]
 
