@@ -141,7 +141,7 @@ public class TransparentPlayerCharacter extends PlayerCharacter
 		public Float getRank(Skill sk)
 		{
 			return ((skillSet == null) || (skillSet.get(sk) == null)) ? 0.0f
-				: new Float(skillSet.get(sk));
+				: Float.valueOf(skillSet.get(sk));
 		}
 
 	}
@@ -195,7 +195,7 @@ public class TransparentPlayerCharacter extends PlayerCharacter
 	public Float getRank(Skill sk)
 	{
 		return ((skillSet == null) || (skillSet.get(sk) == null)) ? 0.0f
-			: new Float(skillSet.get(sk));
+			: Float.valueOf(skillSet.get(sk));
 	}
 
 	@Override
